chore: subgraph fixes
