+++ tools/subgraph-example/package.json
@@ -1,5 +1,6 @@
 {
   "name": "hardhat-project",
+  "type": "commonjs",
   "devDependencies": {
     "@hashgraph/hedera-local": "^2.16.0",
     "@nomiclabs/hardhat-ethers": "^2.2.1",
