+++ tools/subgraph-example/tsconfig.json
@@ -9,5 +9,5 @@
     "resolveJsonModule": true
   },
   "include": ["./scripts", "./test", "./typechain"],
-  "files": ["./hardhat.config.ts"]
+  "files": ["./hardhat.config.js"]
 }
